More aggressive use of move_path_to_trash to fix permissions issues

+++ conda/install.py
@@ -261,16 +261,26 @@ def rm_rf(path, max_retries=5, trash=True):
         # islink('/path/to/dead-link') is True.
         try:
             os.unlink(path)
+            return
         except (OSError, IOError):
             log.warn("Cannot remove, permission denied: {0}".format(path))
+            if trash and move_path_to_trash(path):
+                return
 
     elif isdir(path):
+
+        # On Windows, always move to trash first.
+        if trash and on_win and move_path_to_trash(path, preclean=False):
+            return
+
         try:
             for i in range(max_retries):
                 try:
                     shutil.rmtree(path, ignore_errors=False, onerror=warn_failed_remove)
                     return
                 except OSError as e:
+                    if trash and move_path_to_trash(path):
+                        return
                     msg = "Unable to delete %s\n%s\n" % (path, e)
                     if on_win:
                         try:
@@ -288,16 +298,6 @@ def rm_rf(path, max_retries=5, trash=True):
                         else:
                             if not isdir(path):
                                 return
-
-                        if trash:
-                            try:
-                                move_path_to_trash(path)
-                                if not isdir(path):
-                                    return
-                            except OSError as e2:
-                                raise
-                                msg += "Retry with onerror failed (%s)\n" % e2
-
                     log.debug(msg + "Retrying after %s seconds..." % i)
                     time.sleep(i)
             # Final time. pass exceptions to caller.
@@ -971,6 +971,8 @@ def is_linked(prefix, dist):
 def delete_trash(prefix=None):
     for pkg_dir in pkgs_dirs:
         trash_dir = join(pkg_dir, '.trash')
+        if not isdir(trash_dir):
+            continue
         try:
             log.debug("Trying to delete the trash dir %s" % trash_dir)
             rm_rf(trash_dir, max_retries=1, trash=False)
@@ -989,14 +991,14 @@ def move_to_trash(prefix, f, tempdir=None):
     return move_path_to_trash(join(prefix, f) if f else prefix)
 
 
-def move_path_to_trash(path):
+def move_path_to_trash(path, preclean=True):
     """
     Move a path to the trash
     """
     # Try deleting the trash every time we use it.
-    delete_trash()
+    if preclean:
+        delete_trash()
 
-    from conda.config import pkgs_dirs
     for pkg_dir in pkgs_dirs:
         trash_dir = join(pkg_dir, '.trash')
 
@@ -1009,14 +1011,17 @@ def move_path_to_trash(path):
         trash_file = tempfile.mktemp(dir=trash_dir)
 
         try:
-            shutil.move(path, trash_file)
+            os.rename(path, trash_file)
         except OSError as e:
             log.debug("Could not move %s to %s (%s)" % (path, trash_file, e))
         else:
+            log.debug("Moved to trash: %s" % (path,))
             delete_linked_data_any(path)
-            return trash_file
+            if not preclean:
+                rm_rf(trash_file, max_retries=1, trash=False)
+            return True
 
-    log.debug("Could not move %s to trash" % path)
+    return False
 
 
 def link(prefix, dist, linktype=LINK_HARD, index=None, shortcuts=False):
@@ -1048,17 +1053,7 @@ def link(prefix, dist, linktype=LINK_HARD, index=None, shortcuts=False):
                 os.makedirs(dst_dir)
             if os.path.exists(dst):
                 log.warn("file already exists: %r" % dst)
-                try:
-                    os.unlink(dst)
-                except OSError:
-                    log.error('failed to unlink: %r' % dst)
-                    if on_win:
-                        try:
-                            move_path_to_trash(dst)
-                        except ImportError:
-                            # This shouldn't be an issue in the installer anyway
-                            pass
-
+                rm_rf(dst)
             lt = linktype
             if f in has_prefix_files or f in no_link or islink(src):
                 lt = LINK_COPY
@@ -1122,18 +1117,7 @@ def unlink(prefix, dist):
         for f in meta['files']:
             dst = join(prefix, f)
             dst_dirs1.add(dirname(dst))
-            try:
-                os.unlink(dst)
-            except OSError:  # file might not exist
-                log.debug("could not remove file: '%s'" % dst)
-                if on_win and os.path.exists(join(prefix, f)):
-                    try:
-                        log.debug("moving to trash")
-                        move_path_to_trash(dst)
-                    except ImportError:
-                        # This shouldn't be an issue in the installer anyway
-                        #   but it can potentially happen with importing conda.config
-                        log.debug("cannot import conda.config; probably not an issue")
+            rm_rf(dst)
 
         # remove the meta-file last
         delete_linked_data(prefix, dist, delete=True)
